Created URLToysClone and added loading animation.
I created a class called URLToysClone that's for anything that works
like a clone of URLToys, with the ability to filter down lists of things
with a bunch of simple terse commands. And then I made PRAWToys a
subclass of it. This should help keep my code organized, because it's
now split up between generic code and PRAW-specific code.

Also I have a little loading indicator, now, for all those processes
that take too long to go. I used threads instead of multiprocesses
because the wrapper needs to know which stdout to use, and you can't
send that information between processes because it's not pickleable.